requirements.txt
python-dotenv
ollama
tweepy
schedule
gradio
feedparser